pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "threadsafe-business-logger"
version = "0.3.0"
description = "Une librairie Python fournissant des loggers métier thread-safe avec backend SQLite ou JSON."
requires-python = ">=3.8"
dependencies = [
    "python-dotenv>=1.0.0",
]

# Rend la commande 'make install-dev' et 'pip install -e ".[dev]"' fonctionnelle.
[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]